Minimum working version of getting posts (#8)
* Add Facebook authentication

This adds Facebook authentication by returning the finished access token for the Facebook graph API.

* Add validation function for authentication elements

Currently twitter_auth and facebook_auth use the same way to validate their authentication elements.
Depending on the requirements of other APIs it might make sense to generically validate the elements.

* Change auth token generation to facepy method

Facepy expects a list for authentication, so we need to generate the auth_token differently.
Facepy already has a utility method to generate this token, so we’re using that.

* Switch to dedicated auth element validation method

I’m not sure if the „if…return“ here is the most pythonic way to handle this.
Suggestions welcome on how to change this/make this better.

* Add Facebook to db setup backend

Rough initial table schema. Still need to discuss what fields to add etc.

* Use different table names

Interactions are a better name than reactions, also increased portability of collect_social codebase into this project.

* Switch back to reactions, remove url and media tables

Looking further into the collect_social code, the url and media tables are not necessary and reactions are what will be used in this project.

* Add barebones functionality to get posts

Adds the bare minimum to get posts, but does not do anything with the posts except print them.
Unclear how facepy handles the limit parameter: what if I want *all* of the posts from a page?
date_range will probably need to be passed on to `on_post` to handle saving.

* Fix Facebook authentication

Facebook auth did not pass tests due to missing handling of api version.

* Add correct api version handling

This was missing from the last commit.

* Add Facebook tests

This adds some very basic auth tests. Needed before we can start getting posts etc.

* Add working post download and test

This is a *very* rudimentary version of getting posts and a minimal test for it to work.